Move profile override tests to a dedicated file.
