+++ tests/testsuite/main.rs
@@ -70,6 +70,7 @@ mod path;
 mod plugins;
 mod proc_macro;
 mod profiles;
+mod profile_overrides;
 mod publish;
 mod read_manifest;
 mod registry;
